requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
psycopg2-binary
pgvector
numpy
pickle5
